Remove MyTileMapEditorFrame (can't use JBTabbedPane outside Intellij), rearrange some code
