+++ src/main/kotlin/com/soywiz/korge/intellij/editor/tile/TileMapEditor.kt
@@ -70,10 +70,10 @@ fun Styled<out Container>.createTileMapEditor(
 						val newTilemap = tilemap.data.clone()
                         newTilemap.width = width
                         newTilemap.height = height
-                        for (pattern in newTilemap.patternLayers) {
+                        for (tiles in newTilemap.tileLayers) {
                             val newMap = Bitmap32(width, height)
-                            newMap.put(pattern.map, 0, 0)
-                            pattern.map = newMap
+                            newMap.put(tiles.map, 0, 0)
+                            tiles.map = newMap
                         }
 						history.addAndDo("RESIZE") { redo ->
 							if (redo) {
@@ -438,115 +438,3 @@ fun Styled<out Container>.createTileMapEditor(
 		}
 	}
 }
-
-
-
-class MyTileMapEditorPanel(
-		val tmxFile: VfsFile,
-		val history: HistoryManager = HistoryManager(),
-		val registerHistoryShortcuts: Boolean = true,
-		val projectCtx: ProjectContext? = null,
-		val onSaveXml: (String) -> Unit = {}
-) : JPanel(BorderLayout()) {
-	val tmx = runBlocking { tmxFile.readTiledMap() }
-	init {
-		styled.createTileMapEditor(tmx, history, registerHistoryShortcuts, projectCtx)
-		history.onSave {
-			runBlocking {
-				val xmlString = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + tmx.toXML().toOuterXmlIndented().toString()
-				onSaveXml(xmlString)
-				//tmxFile.writeString(xmlString)
-			}
-		}
-		//history.onChange {
-		history.onAdd {
-			history.save()
-		}
-	}
-/*
-	val realPanel = tileMapEditor.contentPanel
-
-	val mapComponent = MapComponent(tmx)
-
-	var scale: Double
-		get() = mapComponent.scale
-		set(value) = run { mapComponent.scale = value }
-
-	val mapComponentScroll = JBScrollPane(mapComponent).also { scroll ->
-		//scroll.verticalScrollBar.unitIncrement = 16
-	}
-
-	fun updatedSize() {
-		tileMapEditor.leftSplitPane.dividerLocation = 200
-		tileMapEditor.rightSplitPane.dividerLocation = tileMapEditor.rightSplitPane.width - 200
-	}
-
-	val layersController = LayersController(tileMapEditor.layersPane)
-	val propertiesController = PropertiesController(tileMapEditor.propertiesPane)
-
-	init {
-
-		add(realPanel, BorderLayout.CENTER)
-
-		tileMapEditor.mapPanel.add(mapComponentScroll, GridConstraints().also { it.fill = GridConstraints.FILL_BOTH })
-
-		tileMapEditor.zoomInButton.addActionListener { scale *= 1.5 }
-		tileMapEditor.zoomOutButton.addActionListener { scale /= 1.5 }
-
-
-		updatedSize()
-		addComponentListener(object : ComponentAdapter() {
-			override fun componentResized(e: ComponentEvent) {
-				updatedSize()
-			}
-		})
-	}
- */
-}
-
-/*
-class PropertiesController(val panel: PropertiesPane) {
-	var width = 100
-	var height = 100
-	val propertyTable = KorgePropertyTable(KorgePropertyTable.Properties().register(::width,::height)).also {
-		panel.tablePane.add(JScrollPane(it), BorderLayout.CENTER)
-	}
-}
-
-class LayersController(val panel: LayersPane) {
-	init {
-		val menu = JPopupMenu("Menu").apply {
-			add("Tile Layer")
-			add("Object Layer")
-			add("Image Layer")
-		}
-
-		panel.newButton.addActionListener {
-			menu.show(panel.newButton, 0, panel.newButton.height)
-		}
-	}
-}
- */
-
-class MyTileMapEditorFrame(
-	val tmxFile: VfsFile,
-	val projectCtx: ProjectContext? = null
-) : JFrame() {
-	init {
-		contentPane = MyTileMapEditorPanel(tmxFile, projectCtx = projectCtx)
-		pack()
-	}
-
-	companion object {
-		@JvmStatic
-		fun main(args: Array<String>) {
-			val frame = MyTileMapEditorFrame(localCurrentDirVfs["samples/gfx/sample.tmx"])
-			frame.defaultCloseOperation = EXIT_ON_CLOSE
-			frame.setLocationRelativeTo(null)
-			frame.isVisible = true
-		}
-	}
-}
-
-inline fun Bitmap32.anyFixed(callback: (RGBA) -> Boolean): Boolean = (0 until area).any { callback(data[it]) }
-inline fun Bitmap32.allFixed(callback: (RGBA) -> Boolean): Boolean = (0 until area).all { callback(data[it]) }
